refactor($compile): 공유커널 패키지 추가하여 비속어 검증 로직 분리
