+++ src/main/java/kitchenpos/shared/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain;
+package kitchenpos.shared.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
